test(ci): extract shared withTempDir/createCore test helpers
Three test files (cluster-report.test.js, messenger-report.test.js,
fs-utils.test.js) duplicated the same temp-directory and core-mock
helpers. Move them to a single shared/test-utils.js module and let
each test wrap it with its own prefix via a thin inTempDir helper.
