+++ .github/scripts/js/e2e/report/shared/test-utils.js
@@ -1,0 +1,54 @@
+// Copyright 2026 Flant JSC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//     http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+const fs = require("fs");
+const os = require("os");
+const path = require("path");
+
+/**
+ * Creates a mocked GitHub Actions core object for unit tests.
+ *
+ * @returns {{
+ *   info: jest.Mock,
+ *   warning: jest.Mock,
+ *   setOutput: jest.Mock
+ * }} Mocked core object.
+ */
+function createCore() {
+  return {
+    info: jest.fn(),
+    warning: jest.fn(),
+    setOutput: jest.fn(),
+  };
+}
+
+/**
+ * Runs a test body inside a temporary directory and removes it afterwards.
+ *
+ * @template T
+ * @param {string} prefix Temporary directory name prefix.
+ * @param {function(string): (Promise<T>|T)} testFn Test body.
+ * @returns {Promise<T>} Test result.
+ */
+async function withTempDir(prefix, testFn) {
+  const tempDir = fs.mkdtempSync(path.join(os.tmpdir(), `${prefix}-`));
+  try {
+    return await testFn(tempDir);
+  } finally {
+    fs.rmSync(tempDir, { recursive: true, force: true });
+  }
+}
+
+module.exports = {
+  createCore,
+  withTempDir,
+};
